initial commit: basic library framework
This first commit includes the LICENSE and the .gitignore in addition to
the beginnings of the zlua (name tbd) library.

At this stage src/zlua.zig contains a very basic wrapper around the Lua
C API. The goal is to expose every function in a Zig fashion.

The build.zig script contains a single entry point `zig build test` to
run tests.

The intended use of this library will be to include in other projects
and link as a package/library.