add standard library opening functions
Exposes the luaopen_* and luaL_openlibs. Also adds a Ziggified wrapper
function that takes a packed struct as input to specify which libs to
load.

After looking at rlua (a rust lua library) and the lua reference manual,
it seems that these library loading functions have the possibility of
failure. Not sure how to address this yet, but have left a todo.