add lua_arith and test
Adds Lua.arith for mathematical operations on the stack. Includes a test
where each operator is used at least once. The Zig wrapper uses an enum
to ensure no invalid operators are attempted.

A few other functions are included as needed for testing, but will be
tested individually later on.