tests: Lua Buffers
Adds a test to ensure each of the luaL_Buffer functions work. In
experimenting with this, I have found the Buffer code to be very touchy.
If you don't use it correctly things will break badly. Perhaps there is
something that can be added in the Zig layer for safety.

Requires to initialize a Buffer variable before calling init because the
structure keeps a pointer to an internal array so the struct cannot be
passed or returned by value.