tests: add global table and package tests
Adds tests that load the global table (_G) and inspect values in it. In
order to make this simple, the standard library packages are inspected.

This test led to discovering that the Lua.open() function wasn't working
properly as a replacement for openlibs(). Now it properly uses requireF
to load the desired standard library functions.

The various open* functions are still included, but are likely not very
useful in most common situations.