tests: add setFuncs and newLib tests
The luaL_setfuncs function (macro) expects a C array so it can calculate
sizes with sizeof(). In Zig as a function it would be awkward to pass a
null-terminated array of FnReg values.

So we make things easier by just rewriting the auxlib function in Zig,
which has the benefit of accepting a slice rather than an array.

Adds a test to ensure it works.