refactor: change toString to toBytes
The Lua API has the concept of lstrings (length strings). These are used
when the contents may contain zeroes, or when it is useful to know the
length. Because this data is not necessarily a string, ziglua has
renamed these lstring functions to bytes functions.

The toString function returns a zero-terminated many item pointer which
is useful when you know the data is not zero-terminated.