refactor: remove toNumberX and toIntegerX
Removes these functions in favor of toNumber and toInteger. These
remaining functions error if the value cannot be returned as a number.

This makes the API safer, while also being a bit more cumbersome. But it
is better to be explicit than use the old toNumber and not know it
returns 0 silently on fail.