add *Ex variants of pushString and pushBytes
I never care for the return value of this function, so writing _ = gets
tedious. This adds two new versions that return the value, and the
shorter non *Ex functions return void. This makes the common case
easier, but the option to access the return is still available. This
will be applied to other similar functions in the API.