chore: update for Zig 0.10.0
In anticipation for the Zig release coming soon (and the new self-hosted
compiler finally free of bugs that prevented this project from
compiling), this updates the sources for each version of Lua to work in
the new compiler. Mostly trivial changes.

Note that at this point Lua 5.1 doesn't yet compile due to a bug that
was just fixed in Zig, and will be addressed later on.