Remove vendored lua libraries
Use the Zig package manager to download the sources for Lua 5.1 through
Lua 5.4. Luau still needs to be done this way.

The LuaVersion enum strings are changed to remove the _ characters, this
is a breaking change.

Also makes the tests rely on the module rather than the raw source file.
Not sure if this is a good idea yet, just testing some things.

This links the compiled lua library to the ziglua module so it doesn't
need to be linked manually in other projects.