Expose Lua headers with the Lua artifact
Now the dependency.artifact("lua") includes the Lua headers so other
software (like LPeg for example) can link against the built Lua
artifact.

This also exposes the Luau headers, but those may not be as useful with
all of the changes. For example, there is no lauxlib.h header in Luau.