Change Lua.init() to accept a pointer to an Allocator
The Lua VM uses a void pointer to store some user data that is passed to
the custom allocator function.

This modifies Ziglua to require a pointer to be passed to the library
rather than allocating the pointer internally. This is primarily
motivated by clarity. This way makes it obvious that Ziglua diverges
form the norm in Zig code and requires a pointer to an allocator.

An additional benefit is the lua.allocator field is no longer required.
This makes the struct smaller, and removes confusion (why is
lua.allocator not always set?).

Part of #40