Add lua.allocator() function to get the Zig allocator
Adds a function to return the Zig allocator associated with a Lua state.
The function will fail if the Lua state was not created with a Zig
allocator, or using the returned Allocator will segfault.

Correct use of this function is trusted to the programmer. Though in the
normal/default use of Ziglua, this will always be safe.

Closes #40