Add initial LuaJIT support
Adds the latest release of LuaJIT v2.1 and all necessary build file
changes to build as a static library.

This has only been tested on aarch64 macos. Many LuaJIT flags are
missing from the build config. Even if it works, it may not be working
perfectly.

Uses the Lua 51 library bindings, so some additional LuaJIT functions
may be missing. These will be audited and added at a later time.

Part of #19