Make Lua an opqaque
Rather than storing a *LuaState inside the Lua struct, this changes Lua
to be an opaque type that directly represents the memory passed to a Lua
struct.

Internally, this makes a few things slightly less readable.
@ptrCast(lua) and @as(*LuaState, @ptrCast(lua)) are required when
passing data to the c functions.

But externally, this makes more sense. The value returned from
Lua.init() is always a pointer now, and represents the Lua state itself.