Make pushString and pushStringZ always return a string
In Lua 5.1, LuaJIT, and Luau, pushlstring does not return the pushed
string. I think it is better to make this consistent with the later
versions of Lua in the Zig wrapper, so we can use toString to get the
most recently pushed string.

Though this makes an additional API call for the above mentioned
versions, I think the added consistency is worth it.

Going forward I want to remove as much switch (lang) as possible.