Add safety to toNumber and toInteger in Lua 5.1
This makes the behavior of toNumber and toInteger in Lua 5.1 match that
of all other target Lua versions in Ziglua. That is, if the conversion
fails it will return a Zig error rather than returning a 0.

Although this does add an additional internal function call and
condition, this added safety is worth it.