Modify optNumber and similar to return a null
The Lua versions of these macros accept a default value to be
returned when the argument is nil or absent. This is actually limiting
in Zig when we have the orelse operator.

With orelse, a block of code can follow rather than a simple value. This
makes these functions more powerful.