Remove lazy dependencies
I'm not sure the best solution, but #89 raised an issue when ziglua is a
dependency of another package and lazy dependencies are used. Not sure
on the proper solution there, but this should at least solve it for now.
With the downside of more network requests and disk usage required.