Remove newState, newStateLibc, getAllocF, and setAllocF
In preparation for #106, remove these functions that are not needed.
These functions conflict with the original intention of using a Zig
Allocator to handle all the memory allocation. It should not be
necessary to swap out an allocator function or inspect the allocator
function during runtime, and that would more than likely lead to bugs.

Closes #51