Remove pointer from Lua.init allocator parameter
The allocator is not conventional in Zig code. This wasn't causing any
issues from what I can see, but it is still good to follow the
conventions.

The solution is to store a copy of the allocator in an *Allocator stored
on the heap. This is done in the Lua.init function. That allocator is
then retrieved during Lua.deinit by calling c.lua_getallocf to get the
allocator pointer stored on the heap.

Closes #106