fix: remove `Lua.equal` with compile error for lua 5.2, 5.3, 5.4
`lua_equal` is deprecated in Lua versions 5.2 and later; the same
behavior can be achieved with `lua_compare` with the equality
flag. this commit makes using `lua_equal` a compile error when `lang`
is 5.2, 5.3 or 5.4. in fact, it is already a compile error (since the
C header does not expose the function), but making it fail faster with
a helpful replacement suggestion seems useful.