fix(build)!: update to zig 0.14.0 (#136)
* fix(build)!: update to zig 0.14.0

BREAKING: this PR changes the package name from `ziglua` to
`lua_wrapper`. not using `zig` in the package name has been
repeatedly communicated as best practices, analogous to `npm`'s naming
guidelines. for this particular package, this is complicated by the
fact that `ziglua` provides far more than just `lua` packaged for
Zig's build system (which already exists in numerous places), but also
aims to provide ergonomic and powerful Zig abstractions on top of the
Lua C implementation. For this reason, I chose to add "bindings" to
the name.

NB - the "fingerprint" field in build.zig.zon is new in Zig 0.14.0 and
is computed in part based on the name. once we have settled on the
name, if it is not "lua_wrapper" it will be necessary to delete the
fingerprint line and regenerated it.

* fix(ci): update build runners to zig 0.14.0

* package: lua_bindings -> lua_wrapper