fix: avoid std.meta.Tuple
This patch removes the use of `std.meta.Tuple` for compile-time
metaprogramming. It's not clear to me that the convenience it brings
makes the use of `std.meta` worth it. In particular, parsing the
arguments list twice (once for `TypeOfWrap` and once in `wrap`) seems
like an obvious area for improvement.